Wrote code for question 6.
question_6.py:
Wrote the code for q6. Have it load the antenna data and compute EEPs for all pairs theta and phi for the whole hemisphere domain. Then it loads the visibilities and model matrices and estimates gain solutions for the 2 models. Finally, it computes the array patterns for both feeds and the true, AEP and EEP gain solutions and then pllots the 2D station beam.